requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-model-utils==4.5.1
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
